Use bigints for ID fields

+++ priv/repo/migrations/20180816191344_create_scenes_table.exs
@@ -9,8 +9,8 @@ defmodule Ret.Repo.Migrations.CreateScenesTable do
       add(:name, :string, null: false)
       add(:description, :string)
       add(:state, :string, null: false)
-      add(:author_account_id, :integer, null: false)
-      add(:upload_id, :integer, null: false)
+      add(:author_account_id, :bigint, null: false)
+      add(:upload_id, :bigint, null: false)
       add(:attribution_name, :string, null: false)
       add(:attribution_link, :string)
